New Development: No             JIRA Issue: CAS-3654         Ready for Test: Yes      Interface Changes: No What Interface Changed:          Test Programs: test_tflagcmd   Put in Release Notes: No              Module(s): Module Names change impacts.            Description: Catching when the output file to save the parameters 	   already exists.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@17813 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/xmlcasa/scripts/task_tflagcmd.py
@@ -309,11 +309,12 @@ def tflagcmd(
                     writeFlagCmd(vis, myflagd, valid_rows)
             else:
                 # Save to ascii file
-                try:
-                    ffout = open(outfile, 'w')
-                except:
-                    raise Exception, 'Error opening output file ' \
-                        + outfile
+                if os.path.exists(outfile):
+                    raise Exception, 'Output file already exists, ' \
+                                    + outfile
+                                            
+                ffout = open(outfile, 'w')
+
                 try:
                     for cmd in mycmdl:
                         print >> ffout, '%s' % cmd
@@ -359,10 +360,9 @@ def tflagcmd(
             return myflagcmd
         
     except Exception, instance:
+        casalog.post('%s'%instance,'ERROR')
+        raise
 
-                # tflocal.done()
-        print '*** Error ***', instance
-                # raise
 
     # write history
     try:
